Modification:
Simplified the Search Store Screen. Now it will only list locations 
available (currently six for academic reasons). It is using a Spinner,
which is scrollable if we desired to add more stores.

Removed the ShowStore and associated classes.

Added functionality to the help button. Waiting for help text to put in
(user documentation as part of GP3)

Change SearchForStore. Right now, it is literally searching for store.
It will collect the integer from the position of the spinner, and put it
in the intent. This will pass the information to the next activity (now
ProductSearch)

Modified the Manifest to reflect removing Show Store.